fyp-web/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Security-Too==5.1.2
pandas==1.3.5
argon2_cffi
bcrypt
python-dotenv
google-cloud-storage
sqlalchemy